refactor: address PR #704 review — deduplicate EtherCAT device config code
Extract shared useDeviceConfiguration hook and DeviceConfigurationForm
component to eliminate ~1500 lines of duplication between configured-device-row
and device-detail-panel. Also: centralize EnrichDeviceData type, fix
EtherCATSlaveConfig name collision, replace delete mutation with destructuring,
remove dead renderer-side ESI parser, reduce IPC handler boilerplate with
wrapServiceCall helper, and remove unrelated Architecture Survey from CLAUDE.md.
