+++ src/renderer/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './use-compiler'
 export * from './use-debug-composite-key'
+export * from './use-device-configuration'
 export * from './use-store-selectors'
