+++ src/utils/ethercat/esi-parser.ts
@@ -1,223 +1,17 @@
 /**
- * EtherCAT ESI (EtherCAT Slave Information) XML Parser
+ * EtherCAT ESI Channel Utilities
  *
- * Parses ESI XML files following ETG.2000 specification and extracts
- * device information, PDO mappings, and channel configurations.
+ * Functions for working with parsed ESI device data:
+ * - PDO-to-channel conversion for UI
+ * - IEC 61131-3 address generation
+ * - Default channel mapping generation
+ * - Device summary calculation
+ *
+ * XML parsing is handled exclusively by the main-process parser
+ * in src/main/services/esi-service/esi-parser-main.ts.
  */
 
-import type {
-  ESIChannel,
-  ESIDataType,
-  ESIDevice,
-  ESIDeviceType,
-  ESIFile,
-  ESIFMMU,
-  ESIGroup,
-  ESIParseResult,
-  ESIPdo,
-  ESIPdoEntry,
-  ESISyncManager,
-  ESIVendor,
-  EtherCATChannelMapping,
-} from '@root/types/ethercat/esi-types'
-
-/**
- * Parse hex string to number
- * Handles formats: "#x1234", "0x1234", "1234"
- */
-function parseHexValue(value: string | undefined | null): string {
-  if (!value) return '0x0'
-  const cleaned = value.replace(/#x/gi, '0x')
-  return cleaned.startsWith('0x') ? cleaned : `0x${cleaned}`
-}
-
-/**
- * Get text content from an element by tag name
- */
-function getElementText(parent: Element, tagName: string): string | undefined {
-  const element = parent.getElementsByTagName(tagName)[0]
-  return element?.textContent?.trim() || undefined
-}
-
-/**
- * Get attribute value from an element
- */
-function getAttribute(element: Element, attrName: string): string | undefined {
-  return element.getAttribute(attrName) || undefined
-}
-
-/**
- * Parse Vendor information
- */
-function parseVendor(vendorElement: Element): ESIVendor {
-  return {
-    id: parseHexValue(getElementText(vendorElement, 'Id')),
-    name: getElementText(vendorElement, 'Name') || 'Unknown Vendor',
-  }
-}
-
-/**
- * Parse Group information
- */
-function parseGroup(groupElement: Element): ESIGroup {
-  return {
-    type: getElementText(groupElement, 'Type') || '',
-    name: getElementText(groupElement, 'Name') || '',
-    imageUrl: getElementText(groupElement, 'ImageData16x14'),
-    description: getElementText(groupElement, 'Comment'),
-  }
-}
-
-/**
- * Parse FMMU configuration
- */
-function parseFMMU(fmmuElement: Element): ESIFMMU {
-  const text = fmmuElement.textContent?.trim() || 'Outputs'
-  const validTypes: ESIFMMU['type'][] = ['Outputs', 'Inputs', 'MbxState']
-  return {
-    type: validTypes.includes(text as ESIFMMU['type']) ? (text as ESIFMMU['type']) : 'Outputs',
-  }
-}
-
-/**
- * Parse Sync Manager configuration
- */
-function parseSyncManager(smElement: Element, index: number): ESISyncManager {
-  const typeMap: Record<string, ESISyncManager['type']> = {
-    MbxOut: 'MbxOut',
-    MbxIn: 'MbxIn',
-    Outputs: 'Outputs',
-    Inputs: 'Inputs',
-  }
-
-  return {
-    index,
-    startAddress: parseHexValue(getAttribute(smElement, 'StartAddress')),
-    controlByte: parseHexValue(getAttribute(smElement, 'ControlByte')),
-    defaultSize: parseInt(getAttribute(smElement, 'DefaultSize') || '0', 10),
-    enable: getAttribute(smElement, 'Enable') !== '0',
-    type: typeMap[smElement.textContent?.trim() || 'Outputs'] || 'Outputs',
-  }
-}
-
-/**
- * Parse PDO Entry
- */
-function parsePdoEntry(entryElement: Element): ESIPdoEntry | null {
-  const index = getElementText(entryElement, 'Index')
-  const bitLen = parseInt(getElementText(entryElement, 'BitLen') || '0', 10)
-
-  // Skip padding entries (entries without index or with 0 bitlen used for alignment)
-  if (!index && bitLen > 0) {
-    // This is likely a padding entry, we'll include it but mark it
-    return {
-      index: '0x0000',
-      subIndex: '0x00',
-      bitLen,
-      name: 'Padding',
-      dataType: 'BIT',
-    }
-  }
-
-  if (!index) return null
-
-  return {
-    index: parseHexValue(index),
-    subIndex: parseHexValue(getElementText(entryElement, 'SubIndex')),
-    bitLen,
-    name: getElementText(entryElement, 'Name') || 'Unnamed',
-    dataType: getElementText(entryElement, 'DataType') || 'BYTE',
-    comment: getElementText(entryElement, 'Comment'),
-  }
-}
-
-/**
- * Parse PDO (RxPdo or TxPdo)
- */
-function parsePdo(pdoElement: Element): ESIPdo {
-  const entries: ESIPdoEntry[] = []
-  const entryElements = pdoElement.getElementsByTagName('Entry')
-
-  for (let i = 0; i < entryElements.length; i++) {
-    const entry = parsePdoEntry(entryElements[i])
-    if (entry) {
-      entries.push(entry)
-    }
-  }
-
-  return {
-    index: parseHexValue(getElementText(pdoElement, 'Index')),
-    name: getElementText(pdoElement, 'Name') || 'Unnamed PDO',
-    fixed: getAttribute(pdoElement, 'Fixed')?.toLowerCase() === 'true',
-    mandatory: getAttribute(pdoElement, 'Mandatory')?.toLowerCase() === 'true',
-    smIndex: getAttribute(pdoElement, 'Sm') ? parseInt(getAttribute(pdoElement, 'Sm')!, 10) : undefined,
-    entries,
-  }
-}
-
-/**
- * Parse Device Type information
- */
-function parseDeviceType(typeElement: Element): ESIDeviceType {
-  return {
-    productCode: parseHexValue(getAttribute(typeElement, 'ProductCode')),
-    revisionNo: parseHexValue(getAttribute(typeElement, 'RevisionNo')),
-    name: typeElement.textContent?.trim() || 'Unknown Type',
-  }
-}
-
-/**
- * Parse Device
- */
-function parseDevice(deviceElement: Element, groups: ESIGroup[]): ESIDevice {
-  // Parse Type
-  const typeElement = deviceElement.getElementsByTagName('Type')[0]
-  const type = typeElement ? parseDeviceType(typeElement) : { productCode: '0x0', revisionNo: '0x0', name: 'Unknown' }
-
-  // Get group reference
-  const groupType = getElementText(deviceElement, 'GroupType')
-  const group = groups.find((g) => g.type === groupType)
-
-  // Parse FMMUs
-  const fmmu: ESIFMMU[] = []
-  const fmmuElements = deviceElement.getElementsByTagName('Fmmu')
-  for (let i = 0; i < fmmuElements.length; i++) {
-    fmmu.push(parseFMMU(fmmuElements[i]))
-  }
-
-  // Parse Sync Managers
-  const syncManagers: ESISyncManager[] = []
-  const smElements = deviceElement.getElementsByTagName('Sm')
-  for (let i = 0; i < smElements.length; i++) {
-    syncManagers.push(parseSyncManager(smElements[i], i))
-  }
-
-  // Parse RxPDOs
-  const rxPdo: ESIPdo[] = []
-  const rxPdoElements = deviceElement.getElementsByTagName('RxPdo')
-  for (let i = 0; i < rxPdoElements.length; i++) {
-    rxPdo.push(parsePdo(rxPdoElements[i]))
-  }
-
-  // Parse TxPDOs
-  const txPdo: ESIPdo[] = []
-  const txPdoElements = deviceElement.getElementsByTagName('TxPdo')
-  for (let i = 0; i < txPdoElements.length; i++) {
-    txPdo.push(parsePdo(txPdoElements[i]))
-  }
-
-  return {
-    type,
-    name: getElementText(deviceElement, 'Name') || 'Unknown Device',
-    groupName: group?.name,
-    physics: getAttribute(deviceElement, 'Physics'),
-    fmmu,
-    syncManagers,
-    rxPdo,
-    txPdo,
-    description: getElementText(deviceElement, 'Comment'),
-  }
-}
+import type { ESIChannel, ESIDataType, ESIDevice, ESIPdo, EtherCATChannelMapping } from '@root/types/ethercat/esi-types'
 
 /**
  * Map ESI data type to IEC 61131-3 type
@@ -464,105 +258,6 @@ export function generateDefaultChannelMappings(
   }
 
   return mappings
-}
-
-/**
- * Parse ESI XML string
- */
-export function parseESI(xmlString: string, filename?: string): ESIParseResult {
-  const warnings: string[] = []
-
-  try {
-    const parser = new DOMParser()
-    const doc = parser.parseFromString(xmlString, 'text/xml')
-
-    // Check for parse errors
-    const parseError = doc.getElementsByTagName('parsererror')[0]
-    if (parseError) {
-      return {
-        success: false,
-        error: `XML parse error: ${parseError.textContent}`,
-      }
-    }
-
-    // Get root element
-    const root = doc.getElementsByTagName('EtherCATInfo')[0]
-    if (!root) {
-      return {
-        success: false,
-        error: 'Invalid ESI file: Missing EtherCATInfo root element',
-      }
-    }
-
-    // Parse Vendor
-    const vendorElement = root.getElementsByTagName('Vendor')[0]
-    if (!vendorElement) {
-      return {
-        success: false,
-        error: 'Invalid ESI file: Missing Vendor information',
-      }
-    }
-    const vendor = parseVendor(vendorElement)
-
-    // Parse Groups
-    const groups: ESIGroup[] = []
-    const descriptionsElement = root.getElementsByTagName('Descriptions')[0]
-    if (descriptionsElement) {
-      const groupsElement = descriptionsElement.getElementsByTagName('Groups')[0]
-      if (groupsElement) {
-        const groupElements = groupsElement.getElementsByTagName('Group')
-        for (let i = 0; i < groupElements.length; i++) {
-          groups.push(parseGroup(groupElements[i]))
-        }
-      }
-    }
-
-    // Parse Devices
-    const devices: ESIDevice[] = []
-    if (descriptionsElement) {
-      const devicesElement = descriptionsElement.getElementsByTagName('Devices')[0]
-      if (devicesElement) {
-        const deviceElements = devicesElement.getElementsByTagName('Device')
-        for (let i = 0; i < deviceElements.length; i++) {
-          devices.push(parseDevice(deviceElements[i], groups))
-        }
-      }
-    }
-
-    if (devices.length === 0) {
-      warnings.push('No devices found in ESI file')
-    }
-
-    // Parse InfoData (optional metadata)
-    const infoDataElement = root.getElementsByTagName('InfoData')[0]
-    const infoData = infoDataElement
-      ? {
-          version: getElementText(infoDataElement, 'Version'),
-          creationDate: getElementText(infoDataElement, 'CreationDate'),
-          modificationDate: getElementText(infoDataElement, 'ModificationDate'),
-          vendorUrl: getElementText(infoDataElement, 'VendorUrl'),
-        }
-      : undefined
-
-    const result: ESIFile = {
-      vendor,
-      groups,
-      devices,
-      filename,
-      infoData,
-    }
-
-    return {
-      success: true,
-      data: result,
-      warnings: warnings.length > 0 ? warnings : undefined,
-    }
-  } catch (error) {
-    return {
-      success: false,
-      error: `Failed to parse ESI file: ${error instanceof Error ? error.message : String(error)}`,
-    }
-  }
 }
 
 /**
